[airborne] rename arch/omap to arch/linux
