+++ conf/Makefile.ardrone2
@@ -20,10 +20,10 @@
 #
 
 #
-# This Makefile uses the generic Makefile.omap and adds upload rules for the ARDrone2
+# This Makefile uses the generic Makefile.linux and adds upload rules for the ARDrone2
 #
 
-include $(PAPARAZZI_SRC)/conf/Makefile.omap
+include $(PAPARAZZI_SRC)/conf/Makefile.linux
 
 DRONE = $(PAPARAZZI_SRC)/sw/ext/ardrone2_drivers/ardrone2.py
