+++ conf/boards/ardrone2_sdk.makefile
@@ -10,9 +10,9 @@ BOARD_VERSION=2
 BOARD_TYPE=sdk
 BOARD_CFG=\"boards/$(BOARD)$(BOARD_VERSION)_$(BOARD_TYPE).h\"
 
-ARCH=omap
+ARCH=linux
 $(TARGET).ARCHDIR = $(ARCH)
-# include conf/Makefile.ardrone2 (with specific upload rules) instead of only Makefile.omap:
+# include conf/Makefile.ardrone2 (with specific upload rules) instead of only Makefile.linux:
 ap.MAKEFILE = ardrone2
 
 # -----------------------------------------------------------------------
