+++ conf/boards/lisa_l_1.1.makefile
@@ -16,7 +16,7 @@ $(TARGET).LDSCRIPT=$(SRC_ARCH)/lisa-l.ld
 # -----------------------------------------------------------------------
 ifeq ($(BOARD_PROCESSOR),'omap')
 
-	ARCH  = omap
+	ARCH  = linux
 	$(TARGET).LDFLAGS += -levent -lm
 
 # -----------------------------------------------------------------------
